address review comments
- refactor serde module
- rename InvalidNodeID error variant
- better variable names in loops
- reduce error handling code in `load_files`
- delete old serialization code
- serialize Position::containing_line and Position::trimmed_line
